+++ packages/wdio-cli/src/types.ts
@@ -35,6 +35,7 @@ export interface Questionnair {
     createPackageJSON?: boolean
     projectRootCorrect?: boolean
     projectRoot?: string
+    setupMobileEnvironment?: boolean
 }
 
 export interface ParsedAnswers extends Omit<Questionnair, 'runner' | 'framework' | 'reporters' | 'services' | 'plugins'> {